requirements.txt
Flask==2.3.3
RPi.GPIO==0.7.1
gpiozero==1.6.2
requests==2.31.0
